# OpenAPI-Specification.json
{
    "openapi": "3.0.2",
    "info": {
      "title": "League of Legends Backend",
      "version": "1.0",
      "license": {
        "name": "MIT",
        "url": "https://github.com/Knniff/leaguebackend/blob/master/LICENSE.txt"
      },
      "description": "A simple api for getting league mastery stats and summoner data. A nodejs server with a mongodb in the background."
    },
    "servers": [
      {
        "url": "localhost:4000"
      }
    ],
...
    "paths": {
...
      "/league/summoner/name/{summonerName}": {
        "get": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Get the summoner info for a specific name.",
          "tags": [
            "league"
          ],
          "operationId": "summonerName",
          "parameters": [
            {
              "in": "path",
              "name": "summonerName",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "The requested summoner directly from the Riot-API.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "id",
                      "accountId",
                      "puuid",
                      "name",
                      "profileIconId",
                      "revisionDate",
                      "summonerLevel"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "accountId": {
                        "type": "string"
                      },
                      "puuid": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "profileIconId": {
                        "type": "integer"
                      },
                      "revisionDate": {
                        "type": "integer"
                      },
                      "summonerLevel": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/league/summoner/id/{summonerId}": {
        "get": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Get the summoner info for a specific id from the database, when the id is seen for the first time a summoner is added to the DB and the masteries are fetched.",
          "tags": [
            "league"
          ],
          "operationId": "summonerId",
          "parameters": [
            {
              "in": "path",
              "name": "summonerId",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "The updated summoner from the DB.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "_id",
                      "id",
                      "summonerId",
                      "accountId",
                      "puuid",
                      "summonerName",
                      "summonerLevel",
                      "iconId",
                      "serverId",
                      "createdAt",
                      "updatedAt",
                      "__v"
                    ],
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "summonerId": {
                        "type": "string"
                      },
                      "accountId": {
                        "type": "string"
                      },
                      "puuid": {
                        "type": "string"
                      },
                      "summonerName": {
                        "type": "string"
                      },
                      "summonerLevel": {
                        "type": "integer"
                      },
                      "iconId": {
                        "type": "integer"
                      },
                      "serverId": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "__v": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/league/mastery": {
        "get": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Get masteries.",
          "tags": [
            "league"
          ],
          "operationId": "masteryAll",
          "parameters": [
            {
              "in": "query",
              "name": "limit",
              "schema": {
                "type": "integer",
                "maximum": 50,
                "minimum": 1
              }
            },
            {
              "in": "query",
              "name": "next",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "The masteries in an Array.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "mastery",
                      "next"
                    ],
                    "properties": {
                      "mastery": {
                        "type": "array",
                        "items": {
                          "required": [
                            "mastery"
                          ],
                          "properties": {
                            "mastery": {
                              "type": "object",
                              "properties": {
                                "_id": {
                                  "type": "string"
                                },
                                "id": {
                                  "type": "string"
                                },
                                "championId": {
                                  "type": "integer"
                                },
                                "championLevel": {
                                  "type": "integer"
                                },
                                "championPoints": {
                                  "type": "integer"
                                },
                                "lastPlayTime": {
                                  "type": "integer"
                                },
                                "championPointsSinceLastLevel": {
                                  "type": "integer"
                                },
                                "championPointsUntilNextLevel": {
                                  "type": "integer"
                                },
                                "chestGranted": {
                                  "type": "boolean"
                                },
                                "tokensEarned": {
                                  "type": "integer"
                                },
                                "summonerId": {
                                  "type": "string"
                                },
                                "__v": {
                                  "type": "integer"
                                },
                                "createdAt": {
                                  "type": "string"
                                },
                                "updatedAt": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      },
                      "next": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/league/mastery/summoner/{summonerId}": {
        "get": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Get all masteries from the specified summoner.",
          "tags": [
            "league"
          ],
          "operationId": "masteryAllSummoner",
          "parameters": [
            {
              "in": "query",
              "name": "limit",
              "schema": {
                "type": "integer",
                "maximum": 50,
                "minimum": 1
              }
            },
            {
              "in": "query",
              "name": "next",
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "path",
              "name": "summonerId",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "The masteries in an Array.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "mastery",
                      "next"
                    ],
                    "properties": {
                      "mastery": {
                        "type": "array",
                        "items": {
                          "required": [
                            "mastery"
                          ],
                          "properties": {
                            "mastery": {
                              "type": "object",
                              "properties": {
                                "_id": {
                                  "type": "string"
                                },
                                "id": {
                                  "type": "string"
                                },
                                "championId": {
                                  "type": "integer"
                                },
                                "championLevel": {
                                  "type": "integer"
                                },
                                "championPoints": {
                                  "type": "integer"
                                },
                                "lastPlayTime": {
                                  "type": "integer"
                                },
                                "championPointsSinceLastLevel": {
                                  "type": "integer"
                                },
                                "championPointsUntilNextLevel": {
                                  "type": "integer"
                                },
                                "chestGranted": {
                                  "type": "boolean"
                                },
                                "tokensEarned": {
                                  "type": "integer"
                                },
                                "summonerId": {
                                  "type": "string"
                                },
                                "__v": {
                                  "type": "integer"
                                },
                                "createdAt": {
                                  "type": "string"
                                },
                                "updatedAt": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      },
                      "next": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/league/mastery/champion/{championId}": {
        "get": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Get all masteries from the specified summoner.",
          "tags": [
            "league"
          ],
          "operationId": "masteryAllChampion",
          "parameters": [
            {
              "in": "query",
              "name": "limit",
              "schema": {
                "type": "integer",
                "maximum": 50,
                "minimum": 1
              }
            },
            {
              "in": "query",
              "name": "next",
              "schema": {
                "type": "string"
              }
            },
            {
              "in": "path",
              "name": "championId",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "The masteries in an Array.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "mastery",
                      "next"
                    ],
                    "properties": {
                      "mastery": {
                        "type": "array",
                        "items": {
                          "required": [
                            "mastery"
                          ],
                          "properties": {
                            "mastery": {
                              "type": "object",
                              "properties": {
                                "_id": {
                                  "type": "string"
                                },
                                "id": {
                                  "type": "string"
                                },
                                "championId": {
                                  "type": "integer"
                                },
                                "championLevel": {
                                  "type": "integer"
                                },
                                "championPoints": {
                                  "type": "integer"
                                },
                                "lastPlayTime": {
                                  "type": "integer"
                                },
                                "championPointsSinceLastLevel": {
                                  "type": "integer"
                                },
                                "championPointsUntilNextLevel": {
                                  "type": "integer"
                                },
                                "chestGranted": {
                                  "type": "boolean"
                                },
                                "tokensEarned": {
                                  "type": "integer"
                                },
                                "summonerId": {
                                  "type": "string"
                                },
                                "__v": {
                                  "type": "integer"
                                },
                                "createdAt": {
                                  "type": "string"
                                },
                                "updatedAt": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      },
                      "next": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/league/update/{summonerId}": {
        "put": {
          "security": [
            {
              "bearerAuth": [
  
              ]
            }
          ],
          "description": "Update the masteries for the specified summoner.",
          "tags": [
            "league"
          ],
          "operationId": "updateMastery",
          "parameters": [
            {
              "in": "path",
              "name": "summonerId",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Empty",
              "content": {
                "application/json": {
                  "schema": {
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Wrong input.",
              "content": {
                "application/json": {
                  "schema": {
                    "required": [
                      "Error",
                      "message"
                    ],
                    "properties": {
                      "Error": {
                        "type": "string"
                      },
                      "message": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "array"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
...
    }
  }